# Reimplement storage and summarization for performance