--- a/cmd/routesum/args_test.go
+++ b/cmd/routesum/args_test.go
@@ -25,7 +25,7 @@ func TestArgs(t *testing.T) {
 		require.NoError(t, err, "remove temp directory")
 	}()
 	inPath := filepath.Join(tempdir, "in.txt")
-	err = ioutil.WriteFile(inPath, []byte("192.0.2.0\n192.0.2.1\n"), 0644)
+	err = ioutil.WriteFile(inPath, []byte("192.0.2.0\n192.0.2.1\n"), 0o644)
 	require.NoError(t, err, "write to temp input file")
 
 	args, err = _parseArgs([]string{