--- a/go.mod
+++ b/go.mod
@@ -1,5 +1,9 @@
 module github.com/PatrickCronin/routesum
 
-go 1.15
+go 1.16
 
-require github.com/stretchr/testify v1.7.0
+require (
+	github.com/pkg/errors v0.9.1
+	github.com/stretchr/testify v1.7.0
+	inet.af/netaddr v0.0.0-20210721214506-ce7a8ad02cc1
+)