--- a/pkg/routesum/bitslice/bitslice_test.go
+++ b/pkg/routesum/bitslice/bitslice_test.go
@@ -1,0 +1,206 @@
+package bitslice
+
+import (
+	"testing"
+
+	"inet.af/netaddr"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestNewBitSliceFromBytes(t *testing.T) { // nolint: funlen
+	tests := []struct {
+		name     string
+		ip       string
+		expected BitSlice
+	}{
+		{
+			name: "IPv4",
+			ip:   "192.0.2.37",
+			expected: BitSlice([]byte{
+				1, 1, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 1, 0,
+				0, 0, 1, 0, 0, 1, 0, 1,
+			}),
+		},
+		{
+			name: "IPv6-embedded IPv4",
+			ip:   "::ffff:192.0.2.0",
+			expected: BitSlice([]byte{
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				1, 1, 1, 1, 1, 1, 1, 1,
+				1, 1, 1, 1, 1, 1, 1, 1,
+				1, 1, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 1, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+			}),
+		},
+		{
+			name: "IPv6",
+			ip:   "2001:db8::1",
+			expected: BitSlice([]byte{
+				0, 0, 1, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 1,
+				0, 0, 0, 0, 1, 1, 0, 1,
+				1, 0, 1, 1, 1, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 1,
+			}),
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			ip := netaddr.MustParseIP(test.ip)
+			ipBytes, err := ip.MarshalBinary()
+			require.NoError(t, err)
+
+			ipBits, err := NewFromBytes(ipBytes)
+			require.NoError(t, err)
+
+			assert.Equal(t, test.expected, ipBits, "got expected bits")
+		})
+	}
+}
+
+func TestToBytes(t *testing.T) { //nolint: funlen
+	tests := []struct {
+		name     string
+		bitSlice BitSlice
+		numBytes int
+		expected []byte
+	}{
+		{
+			name: "IPv4 IP",
+			bitSlice: BitSlice([]byte{
+				1, 1, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 1, 0,
+				0, 0, 1, 0, 0, 1, 0, 1,
+			}),
+			numBytes: 4,
+			expected: []byte{192, 0, 2, 37},
+		},
+		{
+			name: "IPv4 network",
+			bitSlice: BitSlice([]byte{
+				1, 1, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 1, 0,
+			}),
+			numBytes: 4,
+			expected: []byte{192, 0, 2, 0},
+		},
+		{
+			name: "IPv4-embedded IPv6 IP",
+			bitSlice: BitSlice([]byte{
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				1, 1, 1, 1, 1, 1, 1, 1,
+				1, 1, 1, 1, 1, 1, 1, 1,
+				1, 1, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 1, 0,
+				0, 0, 1, 0, 0, 1, 0, 1,
+			}),
+			numBytes: 16,
+			expected: []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 192, 0, 2, 37},
+		},
+		{
+			name: "IPv4-embedded IPv6 network",
+			bitSlice: BitSlice([]byte{
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				1, 1, 1, 1, 1, 1, 1, 1,
+				1, 1, 1, 1, 1, 1, 1, 1,
+				1, 1, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 1, 0,
+			}),
+			numBytes: 16,
+			expected: []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 192, 0, 2, 0},
+		},
+		{
+			name: "IPv6 IP",
+			bitSlice: BitSlice([]byte{
+				0, 0, 1, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 1,
+				0, 0, 0, 0, 1, 1, 0, 1,
+				1, 0, 1, 1, 1, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 1,
+			}),
+			numBytes: 16,
+			expected: []byte{0x20, 0x1, 0xd, 0xb8, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
+		},
+		{
+			name: "IPv6 network",
+			bitSlice: BitSlice([]byte{
+				0, 0, 1, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 1,
+				0, 0, 0, 0, 1, 1, 0, 1,
+				1, 0, 1, 1, 1, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 0,
+			}),
+			numBytes: 16,
+			expected: []byte{0x20, 0x1, 0xd, 0xb8, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			assert.Equal(t, test.expected, test.bitSlice.ToBytes(test.numBytes), "got expected bytes")
+		})
+	}
+}