--- a/pkg/routesum/errors.go
+++ b/pkg/routesum/errors.go
@@ -2,7 +2,6 @@ package routesum
 
 import (
 	"fmt"
-	"net"
 )
 
 // InvalidInputErr represents an error ingesting or validating input
@@ -13,18 +12,6 @@ type InvalidInputErr struct {
 func newInvalidInputErrFromString(str string) *InvalidInputErr {
 	return &InvalidInputErr{
 		InvalidValue: str,
-	}
-}
-
-func newInvalidInputErrFromNetIP(ip net.IP) *InvalidInputErr {
-	return &InvalidInputErr{
-		InvalidValue: fmt.Sprintf("%#v", ip),
-	}
-}
-
-func newInvalidInputErrFromNetIPNet(net net.IPNet) *InvalidInputErr {
-	return &InvalidInputErr{
-		InvalidValue: fmt.Sprintf("%#v", net),
 	}
 }