This PR changes the fundamental way routes are stored and the route summarization is calculated.

Resolves #9.