@oschwald 2021-07-26
Have you considered just using a specialized binary tree like `mmdbwriter`? I'd guess that it would be faster, use less memory, and it would allow you to output the networks in order.