@PatrickCronin 2021-07-27
Hi @oschwald! I did consider a specialized binary tree. @rafl told me about the existence of radix trees, and this PR implements a radix-ish tree of radix 2 in [rstrie.go](https://github.com/PatrickCronin/routesum/blob/275dad598120ac382e150aac7ea52f9cfdeb83f3/pkg/routesum/rstrie/rstrie.go).  I haven't looked in detail at the structure implemented in mmdbwriter yet--I'll do that when things settle down a bit--but I'm pretty happy with the shape this one is taking. Does mmdbwriter do something better?