@oschwald 2021-07-28
I saw the radix tree implementation. Nice!

The `mmdbwriter` data structure is less sophisticated but fairly well adapted to this particular problem. It eliminates the need to store the bits separately and makes merging networks very simple.