@PatrickCronin 2021-10-19
I created 3 datasets to use for speed comparisons:

* All IPs in the GeoLite2 Country DB
* All IPs in the GeoLite2 City DB
* All IPs in the [DShield API's Intelfeed endpoint](https://www.dshield.org/api/#intelfeed)

See [mmdb-dump-networks](https://github.com/PatrickCronin/mmdb-dump-networks) and [dshield-intelfeed-ips](https://github.com/PatrickCronin/dshield-intelfeed-ips) for utilities created for the purpose of generating these datasets.

I used go `testing`'s built-in benchmarking functions to check the speed and memory performance of the old code vs the new code on these data sets.

| Source                      | Num IPs    | After Summing | Old Code (s)      | Old Code (MB Allocated) | Old Code (Heap MB at Full Structure) | New Code (s) | New Code (MB Allocated) | New Code (Heap MB at Full Structure) |
|-----------------------------|------------|---------------|-------------------|-------------------------|--------------------------------------|--------------|-------------------------|--------------------------------------|
| DShield API's Intelfeed IPs | 58,131     | 41,765        | 63                | 6615.66                 | 3.25                                 | 2.6          | 17.15                   | 8.06                                 |
| GeoLite2 Country Networks   | 1,843,336  | 96,325        | More than 3 hours | DNF*                    | DNF*                                 | 3.1          | 812.64                  | 29.59                                |
| GeoLite2 City Networks      | 15,530,913 | 96,325        | DNT*              | DNT*                    | DNT*                                 | 22.9         | 6637.09                 | 27.59                                |

`* DNF`: Ran for 3 hours and was killed by the testing platform. I could have overrode this even more, but I didn't see the need, as the new code processed the dataset in 3.1 seconds.
`* DNT`: Based on the performance of the old code on the previous dataset, and this being around 8 times larger, I didn't even try.

Based on the above, it's clear this PR greatly improves the speed and memory usage of the routesumming algorithm. #9 contains an illustration that makes performance of the original code look like it degrades exponentially as input increases, however the x-axis scale itself is essentially exponential so it's hard to determine whether it it actually exponential growth or just linear growth. I'll leave that analysis to another day.

The idea posed by @oschwald, i.e. to replace the radix-like trie with a binary tree, to not store any path-related information in the nodes themselves, but rather to derive it from the edges bewteen them, may yet prove to be an additional improvement. However based on the results above, the PR is good enough to merge currently.